requirements.txt
flask
opencv-python-headless
numpy
requests
gunicorn
Pillow
